renamed sample & updated readme
