--- edit-feature-forms/README.md
@@ -1,30 +1,0 @@
-# Edit Feature Forms
-
-Display and edit feature forms.
-
-![Image of edit feature forms](edit-feature-forms.png)
-
-## Use case
-
-TODO.
-
-## How to use the sample
-
-TODO.
-
-## How it works
-
-TODO.
-
-## Relevant API
-
-* ArcGISMap
-* FeatureForm
-
-## Additional information
-
-This sample uses the GeoViewCompose Toolkit module to be able to implement a Composable MapView.
-
-## Tags
-
-basemap, compose, geoviewcompose, jetpack, map, toolkit
