--- edit-feature-forms/README.metadata.json
@@ -1,31 +1,0 @@
-{
-    "category": "Edit and Manage Data",
-    "description": "Display and edit feature forms.",
-    "formal_name": "EditFeatureForms",
-    "ignore": false,
-    "images": [
-        "edit-feature-forms.png"
-    ],
-    "keywords": [
-        "basemap",
-        "compose",
-        "geoviewcompose",
-        "jetpack",
-        "map",
-        "toolkit",
-        "ArcGISMap",
-        "FeatureForm"
-    ],
-    "language": "kotlin",
-    "redirect_from": "",
-    "relevant_apis": [
-        "ArcGISMap",
-        "FeatureForm"
-    ],
-    "snippets": [
-        "src/main/java/com/esri/arcgismaps/sample/editfeatureforms/MainActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/editfeatureforms/components/MapViewModel.kt",
-        "src/main/java/com/esri/arcgismaps/sample/editfeatureforms/screens/MainScreen.kt"
-    ],
-    "title": "Edit Feature Forms"
-}
