+++ edit-features-using-feature-forms/src/main/AndroidManifest.xml
