+++ edit-features-using-feature-forms/src/main/res/mipmap-hdpi/ic_launcher_round.png
