+++ edit-features-using-feature-forms/src/main/res/mipmap-mdpi/ic_launcher.png
