+++ edit-features-using-feature-forms/src/main/res/mipmap-xhdpi/ic_launcher.png
