+++ edit-features-using-feature-forms/src/main/res/mipmap-xxhdpi/ic_launcher.png
