+++ edit-features-using-feature-forms/src/main/res/mipmap-xxxhdpi/ic_launcher.png
