+++ edit-features-using-feature-forms/src/main/res/values/strings.xml
